create high school and location schema
