+++ config/routes.rb
@@ -21,4 +21,6 @@ Rails.application.routes.draw do
     end
 
   end
+
+  mount Pumi::Engine => '/pumi'
 end
