+++ db/csv/high_school.csv
@@ -1,0 +1,5 @@
+code,label,name_en,parent_code
+0101001,Wat Koh High School,Wat Koh High School,10101
+0102001,Samaki High School,Samaki High School,0102
+0201001,Peam Chi Korng High School,Peam Chi Korng High School,0201
+0202001,Srey Santhor High School,Srey Santhor High School,0202
