+++ db/migrate/20190108080254_change_column_in_career_value.rb
@@ -4,7 +4,7 @@ class ChangeColumnInCareerValue < ActiveRecord::Migration[5.1]
     careerTitles = ['អ្នកថែទាំកុមារ', 'គ្រូបង្រៀនជនពិការ', 'អ្នកបណ្ដុះបណ្ដាលវិជ្ជាជីវៈ']
 
     careers = Career.where(name: careerTitles)
-    return if careers.blank?
+    return if careers.empty?
 
     careers.update_all(categorizable_type: 'Characteristic', categorizable_id: category.id)
   end
