+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20190108080254) do
+ActiveRecord::Schema.define(version: 20190211090259) do
 
   create_table "accounts", force: :cascade, options: "ENGINE=InnoDB DEFAULT CHARSET=utf8" do |t|
     t.string "email", default: "", null: false
@@ -103,6 +103,13 @@ ActiveRecord::Schema.define(version: 20190108080254) do
     t.datetime "updated_at", null: false
     t.datetime "created_date"
     t.string "goal"
+  end
+
+  create_table "high_schools", primary_key: "code", id: :string, limit: 7, force: :cascade, options: "ENGINE=InnoDB DEFAULT CHARSET=utf8" do |t|
+    t.string "name_km"
+    t.string "name_en"
+    t.string "location_code"
+    t.index ["code"], name: "index_high_schools_on_code", unique: true
   end
 
   create_table "majors", force: :cascade, options: "ENGINE=InnoDB DEFAULT CHARSET=utf8" do |t|
