+++ docker-compose.yml
@@ -14,7 +14,7 @@ services:
     build:
       context: .
       dockerfile: Dockerfile.dev
-    command: /bin/sh -c "bundle exec rake db:migrate && bundle exec rails s -p 3000 -b '0.0.0.0'"
+    command: /bin/sh -c "bundle install && bundle exec rake db:migrate && bundle exec rails s -p 3000 -b '0.0.0.0'"
     volumes:
       - .:/app
     ports:
