+++ lib/sample/base.rb
@@ -1,0 +1,23 @@
+# frozen_string_literal: true
+
+require 'csv'
+
+module Sample
+  class Base
+    private_class_method
+
+    def self.csv_path
+      Pathname.new(File.join(Dir.pwd, 'db', 'csv'))
+    end
+
+    def self.write_to_file(data, filename)
+      Dir.mkdir('public/db') unless File.exists?('public/db')
+      file_path = Rails.root.join('public', 'db', "#{filename}.json")
+      content = JSON.pretty_generate(data)
+
+      File.open(file_path, 'w') do |f|
+        f.puts(content)
+      end
+    end
+  end
+end
