+++ lib/tasks/sample.rake
@@ -2,9 +2,6 @@
 
 require 'pathname'
 require 'sample'
-require 'trey_visay_sample'
-require 'school_sample'
-require 'career_sample'
 
 if Rails.env.development? || Rails.env.test?
   namespace :sample do
@@ -19,17 +16,27 @@ if Rails.env.development? || Rails.env.test?
 
     desc 'Loads sample data'
     task load: [:clean_db] do
-      TreyVisaySample.load_samples
+      Sample::TreyVisay.load_samples
     end
 
-    desc 'Export school to json'
-    task export_schools: :environment do
-      SchoolSample.export
+    desc 'Export universities to json'
+    task export_universities: :environment do
+      Sample::University.export
     end
 
     desc 'Export careers to json'
     task export_careers: :environment do
-      CareerSample.export
+      Sample::Career.export
+    end
+
+    desc 'Export high schools to json'
+    task export_high_schools: :environment do
+      Sample::HighSchool.export
+    end
+
+    desc 'Export provinces, districts, and commune to json'
+    task export_locations: :environment do
+      Sample::Location.export
     end
   end
 end
